updated the comments for the coral arm io and moved it into the coral folder
